Datasets/evalplus/Java/src/test/java/com/example/HumanEval_89Test.java
package com.example;
import org.junit.Test;
import static org.junit.Assert.assertEquals;
import com.example.HumanEval_89;

public class HumanEval_89Test {
    @Test
    public void testEncrypt_0() {
        assertEquals("lm", HumanEval_89.encrypt("hi"));
    }

    @Test
    public void testEncrypt_1() {
        assertEquals("ewhjklnop", HumanEval_89.encrypt("asdfghjkl"));
    }

    @Test
    public void testEncrypt_2() {
        assertEquals("kj", HumanEval_89.encrypt("gf"));
    }

    @Test
    public void testEncrypt_3() {
        assertEquals("ix", HumanEval_89.encrypt("et"));
    }

    @Test
    public void testEncrypt_4() {
        assertEquals("jeiajeaijeiak", HumanEval_89.encrypt("faewfawefaewg"));
    }

    @Test
    public void testEncrypt_5() {
        assertEquals("lippsqcjvmirh", HumanEval_89.encrypt("hellomyfriend"));
    }

    @Test
    public void testEncrypt_6() {
        assertEquals("hbdhpqrmpjylqmpyjlpmlyjrqpmqryjlpmqryjljygyjl", HumanEval_89.encrypt("dxzdlmnilfuhmilufhlihufnmlimnufhlimnufhfucufh"));
    }

    @Test
    public void testEncrypt_7() {
        assertEquals("e", HumanEval_89.encrypt("a"));
    }

    @Test
    public void testEncrypt_8() {
        assertEquals("efghi", HumanEval_89.encrypt("abcde"));
    }

    @Test
    public void testEncrypt_9() {
        assertEquals("yzabcd", HumanEval_89.encrypt("uvwxyz"));
    }

    @Test
    public void testEncrypt_10() {
        assertEquals("lipps asvph", HumanEval_89.encrypt("hello world"));
    }

    @Test
    public void testEncrypt_11() {
        assertEquals("gteevevyzzdxi1e2f3gwxtevo1efgx23", HumanEval_89.encrypt("cpaararuvvzte1a2b3cstpark1abct23"));
    }

    @Test
    public void testEncrypt_12() {
        assertEquals("", HumanEval_89.encrypt(""));
    }
}
// End of Test
